feat: withCDN:true sampling + deal client in URL (#10)
Modify the sampling algorithm to fetch Pandora Service deal linked to
the selected ProofSet and skip ProofSets not paying for CDN.

Modify the code building the CDN retrieval URL to use the client wallet
address of the deal payer.

Improve the speed of the sampling algorithm by caching the results
of `getNextProofSetId()`, `proofSetLive()` and `getProofSet()` queries.

Signed-off-by: Miroslav Bajtoš <oss@bajtos.net>